+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 TBA
 
+## [1.1.0] - 2022-02-06
+### Added
+- feature called expression_evaluator that evaluates expressions
+- batchTest.txt to test batches of expressions given
+- expression_evaluator.java that evaluates the expression given
+- input accepts input from either the console or a batch file and solves the expression accordingly
+- output formats the solution to be outputted
+### Changed
+- App.java to be able to evaluate expressions
+
 ## [1.0.0] - 2022-02-05
 ### Added
 - feature called cli_documentation to add help documentation for App.java